Work to try to implement symbolic jumps
Moved symId to ExecutionPath, is object-dependent
Changed DUP opcode to copy symbolic input to new symbolic input
Added instruction pointer instead of counter for opcodes. This is due to jumps being addressed by the bytecode address. Also had to change opcode list to dictionary mapped by bytecode address, and added the next bytecode address for each opcode to the element in the list. So if the first instruction is PUSH1, the next bytecode address is 3.
JUMPs. Implemented with symbolic input checking for IsZero. ExecutionPath then splits into another path, which negates (Not) the symbolic input and does the jump.
main function now iterates through list of paths. executionpaths returned upon splitting or completion.